Trainer Module [Radio Button Checked in Edit-delete]

+++ GymManagementDAL/Data/Migrations/20250930182117_InitialCreate.cs
@@ -78,7 +78,7 @@ namespace GymManagementDAL.Data.Migrations
                 {
                     Id = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
-                    Specialties = table.Column<byte>(type: "tinyint", nullable: false),
+                    Specialties = table.Column<int>(type: "int", nullable: false),
                     HireDate = table.Column<DateTime>(type: "datetime2", nullable: false, defaultValueSql: "GETDATE()"),
                     UpdatedAt = table.Column<DateTime>(type: "datetime2", nullable: true),
                     Name = table.Column<string>(type: "varchar(50)", nullable: false),
